refactor(fl/ui): finish per-element split, drop migration guard
Completes the fl/ui/ folder refactor so the new per-element headers
become canonical and the legacy monolithic files are removed. The
FASTLED_ENABLE_UI_FOLDER_REFACTOR macro is gone — the unity build now
picks up every fl/ui/*.cpp.hpp unconditionally.

Root cause of the temp guard: fl/ui/ was never actually compiled. The
fl/_build.cpp.hpp included fl/ui/_build.cpp.hpp, but no build file in
src/fl/build/ included fl/_build.cpp.hpp directly — only fl.cpp (flat,
same-level only) plus per-subdir fl.<x>+.cpp recursive build files.
There was no fl.ui+.cpp, so the new tree was dead code. Without the
guard, ungating the new tree would have produced duplicate definitions
against the still-active fl/ui.cpp.hpp.

Changes:
- Add src/fl/build/fl.ui+.cpp recursive build entry; register in
  ci/lint_cpp/test_unity_build.py EXPECTED_BUILD_FILES.
- Add src/fl/ui/dropdown.{h,cpp.hpp} (UIDropdown + UIDropdownImpl stub).
- Add src/fl/ui/group.{h,cpp.hpp}     (UIGroup + UIGroupImpl stub).
- Replace src/fl/ui.h with a thin aggregator that re-exports every
  fl/ui/<x>.h so the ~48 existing consumers keep working unchanged.
- Delete src/fl/ui.cpp.hpp and src/fl/ui_impl.h (definitions moved
  into the per-element files).
- Remove fl/ui.cpp.hpp from src/fl/_build.cpp.hpp and src/fl/build/fl.cpp.
- Ungate src/fl/ui/_build.cpp.hpp.

Verified: bash lint passes; unity-build linter passes; fl_ui +
audio_url + 59 incremental dependent tests all green.
