+++ ci/lint_cpp/test_unity_build.py
@@ -78,6 +78,7 @@ EXPECTED_BUILD_FILES = {
     "fl/build/fl.stl+.cpp",
     "fl/build/fl.system+.cpp",
     "fl/build/fl.task+.cpp",
+    "fl/build/fl.ui+.cpp",
     "fl/build/fl.video+.cpp",
     "fl/build/platforms+.cpp",
     "fl/build/third_party+.cpp",
