--- src/fl/ui_impl.h
@@ -1,382 +1,0 @@
-#pragma once
-
-#include "fl/audio/audio_input.h"  // For audio::Config  // IWYU pragma: keep
-#include "fl/stl/string.h"
-#include "fl/stl/url.h"
-#include "fl/stl/compiler_control.h"
-#include "platforms/ui_defs.h"
-#include "fl/stl/noexcept.h"
-
-#ifndef FASTLED_HAS_UI_SLIDER
-#define FASTLED_HAS_UI_SLIDER 0
-#endif
-
-#ifndef FASTLED_HAS_UI_BUTTON
-#define FASTLED_HAS_UI_BUTTON 0
-#endif
-
-#ifndef FASTLED_HAS_UI_CHECKBOX
-#define FASTLED_HAS_UI_CHECKBOX 0
-#endif
-
-#ifndef FASTLED_HAS_UI_NUMBER_FIELD
-#define FASTLED_HAS_UI_NUMBER_FIELD 0
-#endif
-
-#ifndef FASTLED_HAS_UI_TITLE
-#define FASTLED_HAS_UI_TITLE 0
-#endif
-
-#ifndef FASTLED_HAS_UI_DESCRIPTION
-#define FASTLED_HAS_UI_DESCRIPTION 0
-#endif
-
-#ifndef FASTLED_HAS_UI_AUDIO
-#define FASTLED_HAS_UI_AUDIO 0
-#endif
-
-#ifndef FASTLED_HAS_UI_DROPDOWN
-#define FASTLED_HAS_UI_DROPDOWN 0
-#endif
-
-namespace fl {
-
-// If the platform is missing ui components, provide stubs.
-
-#if !FASTLED_HAS_UI_SLIDER
-
-class UISliderImpl {
-  public:
-    // If step is -1, it will be calculated as (max - min) / 100
-    UISliderImpl(const char *name, float value = 128.0f, float min = 1,
-                 float max = 255, float step = -1.f)
-        : mValue(value), mMin(fl::min(min, max)), mMax(fl::max(min, max)) {
-        FASTLED_UNUSED(name);
-        FASTLED_UNUSED(step);
-        if (value < min) {
-            mValue = min;
-        }
-        if (value > max) {
-            mValue = max;
-        }
-    }
-    ~UISliderImpl() FL_NOEXCEPT {}
-    float value() const { return mValue; }
-    float getMax() const { return mMax; }
-    float getMin() const { return mMin; }
-    void setValue(float value) { mValue = fl::max(mMin, fl::min(mMax, value)); }
-    operator float() const { return mValue; }
-    operator u8() const { return static_cast<u8>(mValue); }
-    operator u16() const { return static_cast<u16>(mValue); }
-    operator int() const { return static_cast<int>(mValue); }
-    template <typename T> T as() const { return static_cast<T>(mValue); }
-
-    int as_int() const { return static_cast<int>(mValue); }
-    
-    // Stub method for group setting (does nothing on non-WASM platforms)
-    void setGroup(const fl::string& groupName) { FASTLED_UNUSED(groupName); }
-
-    UISliderImpl &operator=(float value) FL_NOEXCEPT {
-        setValue(value);
-        return *this;
-    }
-    UISliderImpl &operator=(int value) FL_NOEXCEPT {
-        setValue(static_cast<float>(value));
-        return *this;
-    }
-
-  private:
-    float mValue;
-    float mMin;
-    float mMax;
-};
-
-// template operator for >= against a jsSliderImpl
-
-#endif
-
-#if !FASTLED_HAS_UI_BUTTON
-
-class UIButtonImpl {
-  public:
-    UIButtonImpl(const char *name) { FASTLED_UNUSED(name); }
-    ~UIButtonImpl() FL_NOEXCEPT {}
-    bool isPressed() const { return false; }
-    bool clicked() const { return false; }
-    int clickedCount() const { return 0; }
-    operator bool() const { return false; }
-    void click() {}
-    fl::string name() const { return mName; }
-    
-    // Stub method for group setting (does nothing on non-WASM platforms)
-    void setGroup(const fl::string& groupName) { FASTLED_UNUSED(groupName); }
-
-  private:
-    fl::string mName;
-};
-
-#endif
-
-#if !FASTLED_HAS_UI_CHECKBOX
-
-class UICheckboxImpl {
-  public:
-    UICheckboxImpl(const char *name, bool value = false) : mValue(value) {
-        FASTLED_UNUSED(name);
-    }
-    ~UICheckboxImpl() FL_NOEXCEPT {}
-    operator bool() const { return mValue; }
-    explicit operator int() const { return mValue ? 1 : 0; }
-    UICheckboxImpl &operator=(bool value) FL_NOEXCEPT {
-        setValue(value);
-        return *this;
-    }
-    UICheckboxImpl &operator=(int value) FL_NOEXCEPT {
-        setValue(value != 0);
-        return *this;
-    }
-    bool value() const { return mValue; }
-    
-    // Stub method for group setting (does nothing on non-WASM platforms)
-    void setGroup(const fl::string& groupName) { FASTLED_UNUSED(groupName); }
-
-  private:
-    void setValue(bool value) { mValue = value; }
-    bool mValue;
-};
-
-#endif
-
-#if !FASTLED_HAS_UI_NUMBER_FIELD
-
-class UINumberFieldImpl {
-  public:
-    UINumberFieldImpl(const char *name, double value, double min = 0,
-                      double max = 100)
-        : mValue(value), mMin(fl::min(min, max)), mMax(fl::max(min, max)) {
-        FASTLED_UNUSED(name);
-    }
-    ~UINumberFieldImpl() FL_NOEXCEPT {}
-    double value() const { return mValue; }
-    void setValue(double value) { mValue = fl::max(mMin, fl::min(mMax, value)); }
-    operator double() const { return mValue; }
-    operator int() const { return static_cast<int>(mValue); }
-    UINumberFieldImpl &operator=(double value) FL_NOEXCEPT {
-        setValue(value);
-        return *this;
-    }
-    UINumberFieldImpl &operator=(int value) FL_NOEXCEPT {
-        setValue(static_cast<double>(value));
-        return *this;
-    }
-    
-    // Stub method for group setting (does nothing on non-WASM platforms)
-    void setGroup(const fl::string& groupName) { FASTLED_UNUSED(groupName); }
-
-  private:
-    double mValue;
-    double mMin;
-    double mMax;
-};
-
-#endif
-
-#if !FASTLED_HAS_UI_TITLE
-
-class UITitleImpl {
-  public:
-    UITitleImpl(const char *name) { FASTLED_UNUSED(name); }
-    ~UITitleImpl() FL_NOEXCEPT {}
-    
-    // Stub method for group setting (does nothing on non-WASM platforms)
-    void setGroup(const fl::string& groupName) { FASTLED_UNUSED(groupName); }
-};
-
-#endif
-
-#if !FASTLED_HAS_UI_DESCRIPTION
-
-class UIDescriptionImpl {
-  public:
-    UIDescriptionImpl(const char *name) { FASTLED_UNUSED(name); }
-    ~UIDescriptionImpl() FL_NOEXCEPT {}
-    
-    // Stub method for group setting (does nothing on non-WASM platforms)
-    void setGroup(const fl::string& groupName) { FASTLED_UNUSED(groupName); }
-};
-
-#endif
-
-#if !FASTLED_HAS_UI_HELP
-class UIHelpImpl {
-  public:
-    UIHelpImpl(const char *markdownContent) : mContent(markdownContent) { FASTLED_UNUSED(markdownContent); }
-    ~UIHelpImpl() FL_NOEXCEPT {}
-    
-    // Stub method for group setting (does nothing on non-WASM platforms)
-    void setGroup(const fl::string& groupName) { FASTLED_UNUSED(groupName); }
-    
-    // Stub method for accessing markdown content
-    const fl::string& markdownContent() const { return mContent; }
-    
-  private:
-    fl::string mContent;
-};
-
-#endif
-
-#if !FASTLED_HAS_UI_AUDIO
-class UIAudioImpl {
-  public:
-    UIAudioImpl(const fl::string& name) { FASTLED_UNUSED(name); }
-    UIAudioImpl(const fl::string& name, const fl::url& url) {
-        FASTLED_UNUSED(name);
-        FASTLED_UNUSED(url);
-    }
-    UIAudioImpl(const fl::string& name, const fl::audio::Config& config) {
-        FASTLED_UNUSED(name);
-        FASTLED_UNUSED(config);
-    }
-    ~UIAudioImpl() FL_NOEXCEPT {}
-
-    audio::Sample next() {
-        FL_WARN("Audio sample not implemented");
-        return audio::Sample();
-    }
-
-    bool hasNext() {
-        FL_WARN("Audio sample not implemented");
-        return false;
-    }
-
-    // Stub method for group setting (does nothing on non-WASM platforms)
-    void setGroup(const fl::string& groupName) { FASTLED_UNUSED(groupName); }
-
-    // Stub: no underlying audio input on non-WASM platforms
-    fl::shared_ptr<audio::IInput> audioInput() { return nullptr; }
-};
-#endif
-
-#if !FASTLED_HAS_UI_DROPDOWN
-class UIDropdownImpl {
-  public:
-    // Constructor with array of options (size determined automatically)
-    template<fl::size N>
-    UIDropdownImpl(const fl::string &name, const fl::string (&options)[N]) 
-        : UIDropdownImpl(name, options, N) {}
-    
-    // Constructor with fl::vector of options
-    UIDropdownImpl(const fl::string &name, const fl::vector<fl::string>& options) 
-        : mSelectedIndex(0) {
-        FASTLED_UNUSED(name);
-        for (fl::size i = 0; i < options.size(); ++i) {
-            mOptions.push_back(options[i]);
-        }
-        if (mOptions.empty()) {
-            mOptions.push_back(fl::string("No options"));
-        }
-    }
-
-    template<typename Iterator>
-    UIDropdownImpl(const fl::string &name, Iterator begin, Iterator end)
-        : mSelectedIndex(0) {
-        FASTLED_UNUSED(name);
-        for (Iterator it = begin; it != end; ++it) {
-            mOptions.push_back(*it);
-        }
-        if (mOptions.empty()) {
-            mOptions.push_back(fl::string("No options"));
-        }
-    }
-
-    // Constructor with fl::span<fl::string>
-    UIDropdownImpl(const fl::string &name, fl::span<fl::string> options) 
-        : mSelectedIndex(0) {
-        FASTLED_UNUSED(name);
-        for (fl::size i = 0; i < options.size(); ++i) {
-            mOptions.push_back(options[i]);
-        }
-        if (mOptions.empty()) {
-            mOptions.push_back(fl::string("No options"));
-        }
-    }
-
-    // Constructor with initializer_list (only available if C++11 support exists)
-    UIDropdownImpl(const fl::string &name, fl::initializer_list<fl::string> options) 
-        : mSelectedIndex(0) {
-        FASTLED_UNUSED(name);
-        for (const auto& option : options) {
-            mOptions.push_back(option);
-        }
-        if (mOptions.empty()) {
-            mOptions.push_back(fl::string("No options"));
-        }
-    }
-
-    ~UIDropdownImpl() FL_NOEXCEPT {}
-    
-    fl::string value() const { 
-        if (mSelectedIndex < mOptions.size()) {
-            return mOptions[mSelectedIndex]; 
-        }
-        return fl::string("Invalid");
-    }
-    
-    int value_int() const { return static_cast<int>(mSelectedIndex); }
-    
-    void setSelectedIndex(int index) { 
-        if (index >= 0 && index < static_cast<int>(mOptions.size())) {
-            mSelectedIndex = static_cast<fl::size>(index);
-        }
-    }
-    
-    fl::size getOptionCount() const { return mOptions.size(); }
-    fl::string getOption(fl::size index) const {
-        if (index < mOptions.size()) {
-            return mOptions[index];
-        }
-        return fl::string("Invalid");
-    }
-    
-    // Stub method for group setting (does nothing on non-WASM platforms)
-    void setGroup(const fl::string& groupName) { FASTLED_UNUSED(groupName); }
-
-  private:
-    // Private constructor with array of options and count (used by template constructor)
-    UIDropdownImpl(const fl::string &name, const fl::string* options, fl::size count) 
-        : mSelectedIndex(0) {
-        FASTLED_UNUSED(name);
-        for (fl::size i = 0; i < count; ++i) {
-            mOptions.push_back(options[i]);
-        }
-        if (mOptions.empty()) {
-            mOptions.push_back(fl::string("No options"));
-        }
-    }
-
-    fl::vector<fl::string> mOptions;
-    fl::size mSelectedIndex;
-};
-#endif
-
-#ifndef FASTLED_HAS_UI_GROUP
-#define FASTLED_HAS_UI_GROUP 0
-#endif
-
-#if !FASTLED_HAS_UI_GROUP
-
-class UIGroupImpl {
-  public:
-    UIGroupImpl(const char *name) FL_NOEXCEPT : mGroupName(name) {
-        FASTLED_UNUSED(name); 
-    }
-    ~UIGroupImpl() FL_NOEXCEPT {}
-    fl::string name() const FL_NOEXCEPT { return mGroupName; }
-
-  private:
-    fl::string mGroupName;
-};
-
-#endif
-
-} // end namespace fl
